+++ db/migrate/20160725131756_add_category_id_to_event.rb
@@ -1,0 +1,5 @@
+class AddCategoryIdToEvent < ActiveRecord::Migration
+  def change
+    add_reference :events, :category, index: true, foreign_key: true
+  end
+end